requirements.txt
scipy>=1.7.1
biopython>=1.79
click>=7